[#107] fix : 도메인별 메소드 적절한 위치로 변경
